feat(action): add explain code action
