+++ src/main/kotlin/com/github/anhnh01/chatc0de/utils/SingletonHolder.kt
@@ -1,0 +1,27 @@
+package com.github.anhnh01.chatc0de.utils
+
+// ref https://stackoverflow.com/questions/35587652/kotlin-thread-safe-native-lazy-singleton-with-parameter
+
+open class SingletonHolder<out T, in A>(creator: (A) -> T) {
+    private var creator: ((A) -> T)? = creator
+    @Volatile private var instance: T? = null
+
+    fun getInstance(arg: A): T {
+        val i = instance
+        if (i != null) {
+            return i
+        }
+
+        return synchronized(this) {
+            val i2 = instance
+            if (i2 != null) {
+                i2
+            } else {
+                val created = creator!!(arg)
+                instance = created
+                creator = null
+                created
+            }
+        }
+    }
+}
